information alerts

+++ download_video.py
@@ -1,7 +1,17 @@
 import os
 from pytube import YouTube
+from tkinter import messagebox as Messagebox
 from convert_to_mp3 import convert
+from popup import popups
 
+
+def check_url(url):
+    flag = False
+    flag = 'youtube' == url[8:13]
+    if not flag:
+       flag = 'youtu' == url[12:17]
+
+    return flag
 
 def rename_file(file):
     # remove spaces
@@ -15,16 +25,21 @@ def rename_file(file):
     return [input_file, output_file]
 
 
-def action(videos):
-    enlace = videos.get()
-    video = YouTube(enlace)
-    descarga = video.streams.get_audio_only()
-    descarga.download()
-    # get file name
-    file = descarga.default_filename
-    # rename file
-    input_file, output_file = rename_file(file)
-    # convert to mp3
-    convert(input_file, output_file)
-    #remove mp4 file
-    os.remove(input_file)
+def action(url):
+    enlace = url.get()
+    if check_url(enlace):
+        video = YouTube(enlace)
+        descarga = video.streams.get_audio_only()
+        descarga.download()
+        popups('Download', 'Downloading')
+        # get file name
+        file = descarga.default_filename
+        # rename file
+        input_file, output_file = rename_file(file)
+        # convert to mp3
+        convert(input_file, output_file)
+        # remove mp4 file
+        os.remove(input_file)
+        popups('Download', 'Successfully downloaded')
+    else:
+        popups('Process', 'It\'s not a youtube url')
